third_party/blink/web_tests/flag-specific/highdpi/fast/forms/color/color-picker-appearance-color-well-up-accelerated-keyboard-navigation-from-bottom-right-corner-expected.txt

This is a testharness.js-based test.
FAIL Color picker: Moving the color well's selection ring using accelerated keyboard movement. assert_equals: expected "189" but got "184"
Harness: the test ran to completion.
